pyproject.toml
[tool.poetry]
name = "liddar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.3"
pandas = "^2.2.0"
xgboost = "^2.0.3"
scikit-learn = "^1.4.0"
scipy = "^1.12.0"
statsmodels = "^0.14.1"
matplotlib = "^3.8.2"
jupyterlab = "^4.0.12"
sportsdataverse = "^0.0.39"
kaggle = "^1.6.6"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
hyperopt = "^0.2.7"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
